DAGTaskOptimizer: Added a project to test traversing a DAG representing tasks of work.
This is the first test which obviously failed. For one tasks/activities
cannot list their subsequent activities, instead activities need to list
their preceding activities so we know when an activity can be visited
(i.e. only when _all_ of its preceding activities are done).

Also the naive traversal in Visit_NextMostExpensive will favor picking
the next most expensive which isn't necessarily the one on the most
expensive path.