CollisionTestDebugWinforms: Added old project to create and test 2D bounding box collision detection.
Its a WinForms app that you click on to place a box and CTRL+Click to
set direction, use the numpadupdown to change travel distance. This was
the test bed to confirm the correct algorithm for collision detection
between 2D boxes for Entmoot.

The swept approach yields incorrect results because it doesn't handle
when the objects are already interpenetrating. The ray approach handles
that and also pulls the pentrating objects apart correctly.